Refactor MiddlewarePipeline
